moved tests to sub-package
