+++ setup.cfg
@@ -13,7 +13,7 @@ tag_prefix = v
 parentdir_prefix = gwosc
 
 [tool:pytest]
-addopts = --verbose -r s
+addopts = -r s
 
 [coverage:run]
 source = gwosc
